Move static inner classes out of StatementResourse
