+++ presto-main/src/main/java/com/facebook/presto/server/protocol/Query.java
@@ -1,0 +1,649 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.facebook.presto.server.protocol;
+
+import com.facebook.presto.Session;
+import com.facebook.presto.client.ClientTypeSignature;
+import com.facebook.presto.client.Column;
+import com.facebook.presto.client.FailureInfo;
+import com.facebook.presto.client.QueryError;
+import com.facebook.presto.client.QueryResults;
+import com.facebook.presto.client.StageStats;
+import com.facebook.presto.client.StatementStats;
+import com.facebook.presto.execution.QueryExecution;
+import com.facebook.presto.execution.QueryInfo;
+import com.facebook.presto.execution.QueryManager;
+import com.facebook.presto.execution.QueryState;
+import com.facebook.presto.execution.QueryStats;
+import com.facebook.presto.execution.StageInfo;
+import com.facebook.presto.execution.TaskInfo;
+import com.facebook.presto.execution.buffer.PagesSerde;
+import com.facebook.presto.execution.buffer.PagesSerdeFactory;
+import com.facebook.presto.execution.buffer.SerializedPage;
+import com.facebook.presto.metadata.SessionPropertyManager;
+import com.facebook.presto.operator.ExchangeClient;
+import com.facebook.presto.server.SessionContext;
+import com.facebook.presto.spi.ConnectorSession;
+import com.facebook.presto.spi.ErrorCode;
+import com.facebook.presto.spi.Page;
+import com.facebook.presto.spi.QueryId;
+import com.facebook.presto.spi.block.Block;
+import com.facebook.presto.spi.block.BlockEncodingSerde;
+import com.facebook.presto.spi.type.StandardTypes;
+import com.facebook.presto.spi.type.Type;
+import com.facebook.presto.spi.type.TypeSignature;
+import com.facebook.presto.transaction.TransactionId;
+import com.google.common.collect.AbstractIterator;
+import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableSet;
+import com.google.common.collect.Iterables;
+import com.google.common.collect.Lists;
+import com.google.common.util.concurrent.Futures;
+import com.google.common.util.concurrent.ListenableFuture;
+import io.airlift.log.Logger;
+import io.airlift.units.DataSize;
+import io.airlift.units.Duration;
+
+import javax.annotation.concurrent.GuardedBy;
+import javax.annotation.concurrent.ThreadSafe;
+import javax.ws.rs.WebApplicationException;
+import javax.ws.rs.core.Response;
+import javax.ws.rs.core.UriInfo;
+
+import java.net.URI;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.OptionalLong;
+import java.util.Set;
+import java.util.concurrent.Executor;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.atomic.AtomicLong;
+
+import static com.facebook.presto.SystemSessionProperties.isExchangeCompressionEnabled;
+import static com.facebook.presto.spi.StandardErrorCode.GENERIC_INTERNAL_ERROR;
+import static com.facebook.presto.util.Failures.toFailure;
+import static com.google.common.base.Preconditions.checkArgument;
+import static com.google.common.util.concurrent.Futures.immediateFuture;
+import static io.airlift.concurrent.MoreFutures.addTimeout;
+import static io.airlift.units.DataSize.Unit.MEGABYTE;
+import static java.lang.String.format;
+import static java.util.Objects.requireNonNull;
+
+@ThreadSafe
+class Query
+{
+    private static final Logger log = Logger.get(Query.class);
+    private static final long DESIRED_RESULT_BYTES = new DataSize(1, MEGABYTE).toBytes();
+
+    private final QueryManager queryManager;
+    private final QueryId queryId;
+
+    @GuardedBy("this")
+    private final ExchangeClient exchangeClient;
+
+    private final Executor resultsProcessorExecutor;
+    private final ScheduledExecutorService timeoutExecutor;
+
+    @GuardedBy("this")
+    private final PagesSerde serde;
+
+    private final AtomicLong resultId = new AtomicLong();
+    private final Session session;
+
+    @GuardedBy("this")
+    private QueryResults lastResult;
+
+    @GuardedBy("this")
+    private String lastResultPath;
+
+    @GuardedBy("this")
+    private List<Column> columns;
+
+    @GuardedBy("this")
+    private List<Type> types;
+
+    @GuardedBy("this")
+    private Optional<String> setCatalog;
+
+    @GuardedBy("this")
+    private Optional<String> setSchema;
+
+    @GuardedBy("this")
+    private Map<String, String> setSessionProperties;
+
+    @GuardedBy("this")
+    private Set<String> resetSessionProperties;
+
+    @GuardedBy("this")
+    private Map<String, String> addedPreparedStatements;
+
+    @GuardedBy("this")
+    private Set<String> deallocatedPreparedStatements;
+
+    @GuardedBy("this")
+    private Optional<TransactionId> startedTransactionId;
+
+    @GuardedBy("this")
+    private boolean clearTransactionId;
+
+    @GuardedBy("this")
+    private Long updateCount;
+
+    public static Query create(
+            SessionContext sessionContext,
+            String query,
+            QueryManager queryManager,
+            SessionPropertyManager sessionPropertyManager,
+            ExchangeClient exchangeClient,
+            Executor dataProcessorExecutor,
+            ScheduledExecutorService timeoutExecutor,
+            BlockEncodingSerde blockEncodingSerde)
+    {
+        Query result = new Query(sessionContext, query, queryManager, sessionPropertyManager, exchangeClient, dataProcessorExecutor, timeoutExecutor, blockEncodingSerde);
+        result.queryManager.addOutputInfoListener(result.queryId, result::setQueryOutputInfo);
+        return result;
+    }
+
+    private Query(
+            SessionContext sessionContext,
+            String query,
+            QueryManager queryManager,
+            SessionPropertyManager sessionPropertyManager,
+            ExchangeClient exchangeClient,
+            Executor resultsProcessorExecutor,
+            ScheduledExecutorService timeoutExecutor,
+            BlockEncodingSerde blockEncodingSerde)
+    {
+        requireNonNull(sessionContext, "sessionContext is null");
+        requireNonNull(query, "query is null");
+        requireNonNull(queryManager, "queryManager is null");
+        requireNonNull(exchangeClient, "exchangeClient is null");
+        requireNonNull(resultsProcessorExecutor, "resultsProcessorExecutor is null");
+        requireNonNull(timeoutExecutor, "timeoutExecutor is null");
+
+        this.queryManager = queryManager;
+
+        QueryInfo queryInfo = queryManager.createQuery(sessionContext, query);
+        queryId = queryInfo.getQueryId();
+        session = queryInfo.getSession().toSession(sessionPropertyManager);
+        this.exchangeClient = exchangeClient;
+        this.resultsProcessorExecutor = resultsProcessorExecutor;
+        this.timeoutExecutor = timeoutExecutor;
+        requireNonNull(blockEncodingSerde, "serde is null");
+        this.serde = new PagesSerdeFactory(blockEncodingSerde, isExchangeCompressionEnabled(session)).createPagesSerde();
+    }
+
+    public void cancel()
+    {
+        queryManager.cancelQuery(queryId);
+        dispose();
+    }
+
+    public synchronized void dispose()
+    {
+        exchangeClient.close();
+    }
+
+    public QueryId getQueryId()
+    {
+        return queryId;
+    }
+
+    public synchronized Optional<String> getSetCatalog()
+    {
+        return setCatalog;
+    }
+
+    public synchronized Optional<String> getSetSchema()
+    {
+        return setSchema;
+    }
+
+    public synchronized Map<String, String> getSetSessionProperties()
+    {
+        return setSessionProperties;
+    }
+
+    public synchronized Set<String> getResetSessionProperties()
+    {
+        return resetSessionProperties;
+    }
+
+    public synchronized Map<String, String> getAddedPreparedStatements()
+    {
+        return addedPreparedStatements;
+    }
+
+    public synchronized Set<String> getDeallocatedPreparedStatements()
+    {
+        return deallocatedPreparedStatements;
+    }
+
+    public synchronized Optional<TransactionId> getStartedTransactionId()
+    {
+        return startedTransactionId;
+    }
+
+    public synchronized boolean isClearTransactionId()
+    {
+        return clearTransactionId;
+    }
+
+    public synchronized ListenableFuture<QueryResults> waitForResults(OptionalLong token, UriInfo uriInfo, Duration wait)
+    {
+        // before waiting, check if this request has already been processed and cached
+        if (token.isPresent()) {
+            Optional<QueryResults> cachedResult = getCachedResult(token.getAsLong(), uriInfo);
+            if (cachedResult.isPresent()) {
+                return immediateFuture(cachedResult.get());
+            }
+        }
+
+        // wait for a results data or query to finish, up to the wait timeout
+        ListenableFuture<?> futureStateChange = addTimeout(
+                getFutureStateChange(),
+                () -> null,
+                wait,
+                timeoutExecutor);
+
+        // when state changes, fetch the next result
+        return Futures.transform(futureStateChange, ignored -> getNextResult(token, uriInfo), resultsProcessorExecutor);
+    }
+
+    private synchronized ListenableFuture<?> getFutureStateChange()
+    {
+        // if the exchange client is open, wait for data
+        if (!exchangeClient.isClosed()) {
+            return exchangeClient.isBlocked();
+        }
+
+        // otherwise, wait for the query to finish
+        queryManager.recordHeartbeat(queryId);
+        return queryManager.getQueryState(queryId).map(this::queryDoneFuture)
+                .orElse(immediateFuture(null));
+    }
+
+    private synchronized Optional<QueryResults> getCachedResult(long token, UriInfo uriInfo)
+    {
+        // is the a repeated request for the last results?
+        String requestedPath = uriInfo.getAbsolutePath().getPath();
+        if (lastResultPath != null && requestedPath.equals(lastResultPath)) {
+            // tell query manager we are still interested in the query
+            queryManager.getQueryInfo(queryId);
+            queryManager.recordHeartbeat(queryId);
+            return Optional.of(lastResult);
+        }
+
+        if (token < resultId.get()) {
+            throw new WebApplicationException(Response.Status.GONE);
+        }
+
+        // if this is not a request for the next results, return not found
+        if (lastResult.getNextUri() == null || !requestedPath.equals(lastResult.getNextUri().getPath())) {
+            // unknown token
+            throw new WebApplicationException(Response.Status.NOT_FOUND);
+        }
+
+        return Optional.empty();
+    }
+
+    private synchronized QueryResults getNextResult(OptionalLong token, UriInfo uriInfo)
+    {
+        // check if the result for the token have already been created
+        if (token.isPresent()) {
+            Optional<QueryResults> cachedResult = getCachedResult(token.getAsLong(), uriInfo);
+            if (cachedResult.isPresent()) {
+                return cachedResult.get();
+            }
+        }
+
+        // Remove as many pages as possible from the exchange until just greater than DESIRED_RESULT_BYTES
+        // NOTE: it is critical that query results are created for the pages removed from the exchange
+        // client while holding the lock because the query may transition to the finished state when the
+        // last page is removed.  If another thread observes this state before the response is cached
+        // the pages will be lost.
+        ImmutableList.Builder<RowIterable> pages = ImmutableList.builder();
+        long bytes = 0;
+        long rows = 0;
+        while (bytes < DESIRED_RESULT_BYTES) {
+            SerializedPage serializedPage = exchangeClient.pollPage();
+            if (serializedPage == null) {
+                break;
+            }
+
+            Page page = serde.deserialize(serializedPage);
+            bytes += page.getSizeInBytes();
+            rows += page.getPositionCount();
+            pages.add(new RowIterable(session.toConnectorSession(), types, page));
+        }
+
+        // client implementations do not properly handle empty list of data
+        Iterable<List<Object>> data = (rows == 0) ? null : Iterables.concat(pages.build());
+
+        // get the query info before returning
+        // force update if query manager is closed
+        QueryInfo queryInfo = queryManager.getQueryInfo(queryId);
+        queryManager.recordHeartbeat(queryId);
+
+        // TODO: figure out a better way to do this
+        // grab the update count for non-queries
+        if ((data != null) && (queryInfo.getUpdateType() != null) && (updateCount == null) &&
+                (columns.size() == 1) && (columns.get(0).getType().equals(StandardTypes.BIGINT))) {
+            Iterator<List<Object>> iterator = data.iterator();
+            if (iterator.hasNext()) {
+                Number number = (Number) iterator.next().get(0);
+                if (number != null) {
+                    updateCount = number.longValue();
+                }
+            }
+        }
+
+        // close exchange client if the query has failed
+        if (queryInfo.getState().isDone()) {
+            if (queryInfo.getState() == QueryState.FAILED) {
+                exchangeClient.close();
+            }
+            else if (!queryInfo.getOutputStage().isPresent()) {
+                // For simple executions (e.g. drop table), there will never be an output stage,
+                // so close the exchange as soon as the query is done.
+                exchangeClient.close();
+
+                // Return a single value for clients that require a result.
+                columns = ImmutableList.of(new Column("result", "boolean", new ClientTypeSignature(StandardTypes.BOOLEAN, ImmutableList.of())));
+                data = ImmutableSet.of(ImmutableList.of(true));
+            }
+        }
+
+        // only return a next if the query is not done or there is more data to send (due to buffering)
+        URI nextResultsUri = null;
+        if (!queryInfo.isFinalQueryInfo() || !exchangeClient.isClosed()) {
+            nextResultsUri = createNextResultsUri(uriInfo);
+        }
+
+        // update catalog and schema
+        setCatalog = queryInfo.getSetCatalog();
+        setSchema = queryInfo.getSetSchema();
+
+        // update setSessionProperties
+        setSessionProperties = queryInfo.getSetSessionProperties();
+        resetSessionProperties = queryInfo.getResetSessionProperties();
+
+        // update preparedStatements
+        addedPreparedStatements = queryInfo.getAddedPreparedStatements();
+        deallocatedPreparedStatements = queryInfo.getDeallocatedPreparedStatements();
+
+        // update startedTransactionId
+        startedTransactionId = queryInfo.getStartedTransactionId();
+        clearTransactionId = queryInfo.isClearTransactionId();
+
+        // first time through, self is null
+        QueryResults queryResults = new QueryResults(
+                queryId.toString(),
+                uriInfo.getRequestUriBuilder().replaceQuery(queryId.toString()).replacePath("query.html").build(),
+                findCancelableLeafStage(queryInfo),
+                nextResultsUri,
+                columns,
+                data,
+                toStatementStats(queryInfo),
+                toQueryError(queryInfo),
+                queryInfo.getUpdateType(),
+                updateCount);
+
+        // cache the last results
+        if (lastResult != null && lastResult.getNextUri() != null) {
+            lastResultPath = lastResult.getNextUri().getPath();
+        }
+        else {
+            lastResultPath = null;
+        }
+        lastResult = queryResults;
+        return queryResults;
+    }
+
+    private synchronized void setQueryOutputInfo(QueryExecution.QueryOutputInfo outputInfo)
+    {
+        // if first callback, set column names
+        if (columns == null) {
+            List<String> columnNames = outputInfo.getColumnNames();
+            List<Type> columnTypes = outputInfo.getColumnTypes();
+            checkArgument(columnNames.size() == columnTypes.size(), "Column names and types size mismatch");
+
+            ImmutableList.Builder<Column> list = ImmutableList.builder();
+            for (int i = 0; i < columnNames.size(); i++) {
+                String name = columnNames.get(i);
+                TypeSignature typeSignature = columnTypes.get(i).getTypeSignature();
+                String type = typeSignature.toString();
+                list.add(new Column(name, type, new ClientTypeSignature(typeSignature)));
+            }
+            columns = list.build();
+            types = outputInfo.getColumnTypes();
+        }
+
+        for (URI outputLocation : outputInfo.getBufferLocations()) {
+            exchangeClient.addLocation(outputLocation);
+        }
+        if (outputInfo.isNoMoreBufferLocations()) {
+            exchangeClient.noMoreLocations();
+        }
+    }
+
+    private ListenableFuture<?> queryDoneFuture(QueryState currentState)
+    {
+        if (currentState.isDone()) {
+            return immediateFuture(null);
+        }
+        return Futures.transformAsync(queryManager.getStateChange(queryId, currentState), this::queryDoneFuture);
+    }
+
+    private synchronized URI createNextResultsUri(UriInfo uriInfo)
+    {
+        return uriInfo.getBaseUriBuilder().replacePath("/v1/statement").path(queryId.toString()).path(String.valueOf(resultId.incrementAndGet())).replaceQuery("").build();
+    }
+
+    private static StatementStats toStatementStats(QueryInfo queryInfo)
+    {
+        QueryStats queryStats = queryInfo.getQueryStats();
+        StageInfo outputStage = queryInfo.getOutputStage().orElse(null);
+
+        return StatementStats.builder()
+                .setState(queryInfo.getState().toString())
+                .setQueued(queryInfo.getState() == QueryState.QUEUED)
+                .setScheduled(queryInfo.isScheduled())
+                .setNodes(globalUniqueNodes(outputStage).size())
+                .setTotalSplits(queryStats.getTotalDrivers())
+                .setQueuedSplits(queryStats.getQueuedDrivers())
+                .setRunningSplits(queryStats.getRunningDrivers() + queryStats.getBlockedDrivers())
+                .setCompletedSplits(queryStats.getCompletedDrivers())
+                .setUserTimeMillis(queryStats.getTotalUserTime().toMillis())
+                .setCpuTimeMillis(queryStats.getTotalCpuTime().toMillis())
+                .setWallTimeMillis(queryStats.getTotalScheduledTime().toMillis())
+                .setQueuedTimeMillis(queryStats.getQueuedTime().toMillis())
+                .setElapsedTimeMillis(queryStats.getElapsedTime().toMillis())
+                .setProcessedRows(queryStats.getRawInputPositions())
+                .setProcessedBytes(queryStats.getRawInputDataSize().toBytes())
+                .setRootStage(toStageStats(outputStage))
+                .build();
+    }
+
+    private static StageStats toStageStats(StageInfo stageInfo)
+    {
+        if (stageInfo == null) {
+            return null;
+        }
+
+        com.facebook.presto.execution.StageStats stageStats = stageInfo.getStageStats();
+
+        ImmutableList.Builder<StageStats> subStages = ImmutableList.builder();
+        for (StageInfo subStage : stageInfo.getSubStages()) {
+            subStages.add(toStageStats(subStage));
+        }
+
+        Set<String> uniqueNodes = new HashSet<>();
+        for (TaskInfo task : stageInfo.getTasks()) {
+            // todo add nodeId to TaskInfo
+            URI uri = task.getTaskStatus().getSelf();
+            uniqueNodes.add(uri.getHost() + ":" + uri.getPort());
+        }
+
+        return StageStats.builder()
+                .setStageId(String.valueOf(stageInfo.getStageId().getId()))
+                .setState(stageInfo.getState().toString())
+                .setDone(stageInfo.getState().isDone())
+                .setNodes(uniqueNodes.size())
+                .setTotalSplits(stageStats.getTotalDrivers())
+                .setQueuedSplits(stageStats.getQueuedDrivers())
+                .setRunningSplits(stageStats.getRunningDrivers() + stageStats.getBlockedDrivers())
+                .setCompletedSplits(stageStats.getCompletedDrivers())
+                .setUserTimeMillis(stageStats.getTotalUserTime().toMillis())
+                .setCpuTimeMillis(stageStats.getTotalCpuTime().toMillis())
+                .setWallTimeMillis(stageStats.getTotalScheduledTime().toMillis())
+                .setProcessedRows(stageStats.getRawInputPositions())
+                .setProcessedBytes(stageStats.getRawInputDataSize().toBytes())
+                .setSubStages(subStages.build())
+                .build();
+    }
+
+    private static Set<String> globalUniqueNodes(StageInfo stageInfo)
+    {
+        if (stageInfo == null) {
+            return ImmutableSet.of();
+        }
+        ImmutableSet.Builder<String> nodes = ImmutableSet.builder();
+        for (TaskInfo task : stageInfo.getTasks()) {
+            // todo add nodeId to TaskInfo
+            URI uri = task.getTaskStatus().getSelf();
+            nodes.add(uri.getHost() + ":" + uri.getPort());
+        }
+
+        for (StageInfo subStage : stageInfo.getSubStages()) {
+            nodes.addAll(globalUniqueNodes(subStage));
+        }
+        return nodes.build();
+    }
+
+    private static URI findCancelableLeafStage(QueryInfo queryInfo)
+    {
+        // if query is running, find the leaf-most running stage
+        return queryInfo.getOutputStage().map(Query::findCancelableLeafStage).orElse(null);
+    }
+
+    private static URI findCancelableLeafStage(StageInfo stage)
+    {
+        // if this stage is already done, we can't cancel it
+        if (stage.getState().isDone()) {
+            return null;
+        }
+
+        // attempt to find a cancelable sub stage
+        // check in reverse order since build side of a join will be later in the list
+        for (StageInfo subStage : Lists.reverse(stage.getSubStages())) {
+            URI leafStage = findCancelableLeafStage(subStage);
+            if (leafStage != null) {
+                return leafStage;
+            }
+        }
+
+        // no matching sub stage, so return this stage
+        return stage.getSelf();
+    }
+
+    private static QueryError toQueryError(QueryInfo queryInfo)
+    {
+        FailureInfo failure = queryInfo.getFailureInfo();
+        if (failure == null) {
+            QueryState state = queryInfo.getState();
+            if ((!state.isDone()) || (state == QueryState.FINISHED)) {
+                return null;
+            }
+            log.warn("Query %s in state %s has no failure info", queryInfo.getQueryId(), state);
+            failure = toFailure(new RuntimeException(format("Query is %s (reason unknown)", state))).toFailureInfo();
+        }
+
+        ErrorCode errorCode;
+        if (queryInfo.getErrorCode() != null) {
+            errorCode = queryInfo.getErrorCode();
+        }
+        else {
+            errorCode = GENERIC_INTERNAL_ERROR.toErrorCode();
+            log.warn("Failed query %s has no error code", queryInfo.getQueryId());
+        }
+        return new QueryError(
+                failure.getMessage(),
+                null,
+                errorCode.getCode(),
+                errorCode.getName(),
+                errorCode.getType().toString(),
+                failure.getErrorLocation(),
+                failure);
+    }
+
+    private static class RowIterable
+            implements Iterable<List<Object>>
+    {
+        private final ConnectorSession session;
+        private final List<Type> types;
+        private final Page page;
+
+        private RowIterable(ConnectorSession session, List<Type> types, Page page)
+        {
+            this.session = session;
+            this.types = ImmutableList.copyOf(requireNonNull(types, "types is null"));
+            this.page = requireNonNull(page, "page is null");
+        }
+
+        @Override
+        public Iterator<List<Object>> iterator()
+        {
+            return new RowIterator(session, types, page);
+        }
+    }
+
+    private static class RowIterator
+            extends AbstractIterator<List<Object>>
+    {
+        private final ConnectorSession session;
+        private final List<Type> types;
+        private final Page page;
+        private int position = -1;
+
+        private RowIterator(ConnectorSession session, List<Type> types, Page page)
+        {
+            this.session = session;
+            this.types = types;
+            this.page = page;
+        }
+
+        @Override
+        protected List<Object> computeNext()
+        {
+            position++;
+            if (position >= page.getPositionCount()) {
+                return endOfData();
+            }
+
+            List<Object> values = new ArrayList<>(page.getChannelCount());
+            for (int channel = 0; channel < page.getChannelCount(); channel++) {
+                Type type = types.get(channel);
+                Block block = page.getBlock(channel);
+                values.add(type.getObjectValue(session, block, position));
+            }
+            return Collections.unmodifiableList(values);
+        }
+    }
+}
